+++ CHANGELOG.md
@@ -13,6 +13,10 @@
 * Should save the latest text input if you close the page quickly (#1067)
 * Keyboard won't show up in code snippet on Android 7 (#1024)
 
+### Middleware ⚙
+
+* Updated middleware protocol to `0.13.24` (#1100)
+
 ## Version 0.1.2
 
 ### New features & enhancements 🚀